refactor: move the package of integration test at araongo properly
