+++ jnosql-arangodb/src/test/java/org/eclipse/jnosql/databases/arangodb/integration/MainStepType.java
@@ -12,7 +12,7 @@
  *
  *    Otavio Santana
  */
-package org.eclipse.jnosql.databases.arangodb.mapping;
+package org.eclipse.jnosql.databases.arangodb.integration;
 
 public enum MainStepType {
     MAIN, ENTRY
